Fold Graph_yaml_importer.hpp into Graph.hpp
This allows Graph to use the yaml importer without having a cyclic
dependency between headers, and without moving stuff into .cpp files.

Part of task #246
