+++ core/CMakeLists.txt
@@ -11,7 +11,6 @@ set(public_headers
     include/gnuradio-4.0/CircularBuffer.hpp
     include/gnuradio-4.0/ClaimStrategy.hpp
     include/gnuradio-4.0/DataSet.hpp
-    include/gnuradio-4.0/Graph_yaml_importer.hpp
     include/gnuradio-4.0/Graph.hpp
     include/gnuradio-4.0/HistoryBuffer.hpp
     include/gnuradio-4.0/LifeCycle.hpp
