+++ core/test/qa_Settings.cpp
@@ -7,7 +7,6 @@
 #include <gnuradio-4.0/Block.hpp>
 #include <gnuradio-4.0/Buffer.hpp>
 #include <gnuradio-4.0/Graph.hpp>
-#include <gnuradio-4.0/Graph_yaml_importer.hpp>
 #include <gnuradio-4.0/Scheduler.hpp>
 #include <gnuradio-4.0/Settings.hpp>
 #include <gnuradio-4.0/Tag.hpp>
